clear code of RaportView
